♻️ refactor: postLikesGathering 함수 분리
