+++ src/pages/likes/components/LikesGatheringsList.tsx
@@ -1,49 +1,16 @@
-import { useInfiniteQuery, QueryFunctionContext } from '@tanstack/react-query';
+import { useInfiniteQuery } from '@tanstack/react-query';
 import useInfiniteScroll from '@/hooks/useInfiniteScroll';
 import Null from '@/components/common/Null';
 import Card from '@/components/card/gathering/Card';
 import { GatheringList } from '@/types';
-import apiRequest from '@/utils/apiRequest';
-import { MainType } from '@/constants/MainList';
-import { getLikes } from '@/utils/likesgathering';
-
-interface likesGatheringsProps {
-  mainType: MainType;
-  subType: string;
-}
-
-interface RequestData {
-  gatheringIds: number[];
-}
+import postLikesGatherings, {
+  likesGatheringsProps,
+} from '../service/postLikesGathering';
 
 export default function LikesGatheringsList({
   mainType,
   subType,
 }: likesGatheringsProps) {
-  // 한 페이지당 모임 수
-  const ROWS_PER_PAGE = 8;
-
-  const postlikesGatherings = async ({
-    pageParam = 0,
-  }: QueryFunctionContext) => {
-    const apiEndpoint = '/api/v1/gatherings/likes';
-    const queryParams = {
-      sortBy: 'deadline',
-      sortDirection: 'ASC',
-      page: String(pageParam),
-      pageSize: String(ROWS_PER_PAGE),
-      ...(mainType !== '전체' && { mainType }),
-      ...(subType !== '전체' && { subType }),
-    };
-
-    const param = `${apiEndpoint}?${new URLSearchParams(queryParams).toString()}`;
-    return await apiRequest<GatheringList, RequestData>({
-      param,
-      method: 'post',
-      requestData: { gatheringIds: getLikes() },
-    });
-  };
-
   // React Query를 사용한 무한 스크롤 데이터 처리
   const {
     data,
@@ -54,7 +21,7 @@ export default function LikesGatheringsList({
     error,
   } = useInfiniteQuery<GatheringList, Error>({
     queryKey: ['likesGatherings', mainType, subType],
-    queryFn: postlikesGatherings,
+    queryFn: (context) => postLikesGatherings({ mainType, subType }, context),
     getNextPageParam: (lastPage, allPages) => {
       return lastPage.content.length > 0 ? allPages.length : undefined;
     },
